LayoutTests/media/video-defaultplaybackrate-before-load-expected.txt


RUN(video.defaultPlaybackRate = 0.5)
EVENT(canplaythrough)
EXPECTED (video.playbackRate == '0.5') OK
EXPECTED (video.defaultPlaybackRate == '0.5') OK
EVENT(timeupdate)
EVENT(timeupdate)
EXPECTED (effectivePlaybackRate < '0.75') OK
END OF TEST
